Add the ability to use a mixin file for a css file.
To use features like @apply in css, we need to be able to load a file that contains the mixins.

When that file is mixed in, the utility classes become available to the css file that is being processed.

````
Rails.application.configure do
  path = File.join(Rails.root, "app", "assets")

  Deadfire.configure do |deadfire_config|
    deadfire_config.asset_registry.register_path(File.join(path, "test.css"), File.join(path, "vendor", "vendor.css"))
  end
end
```

Example configuration allows you to mixin vendor.css into the test.css file.

The mixins are cached, in propshaft one asset file is processed at a time and to speed up mixin files
they are cached.